Create extract_matched_reads_Cloud_v3.0.sh
This script takes a list of SRA accession numbers and a fasta file containing a sequence of interest. For each accession, bbduk is used to collect reads that share at least one 25-mer with the sequence of interest. The 'matching' reads are saved to a fastq file for downstream processing.